Added tokens strtok strspn strcspn isletter

+++ scilab/modules/string/sci_gateway/cpp/sci_strchr.cpp
@@ -83,7 +83,7 @@ types::Function::ReturnValue sci_strchr(types::typed_list &in, int _iRetCount, t
         {
             j = i; /* Input parameter One & two have same dimension */
         }
-        int iLen = wcslen(pCharSample->get(j));
+        int iLen = (int)wcslen(pCharSample->get(j));
         if(iLen != 1)
         {
     		ScierrorW(999,_W("%ls: Wrong type for input argument #%d: Char(s) expected.\n"),L"strchr", 2);
